comment: update how passing block type to comment
i think i fixed a loophole by setting the comment selection type using
the block's type instead of calling thread['selection_type'], as the
latter value is set in the thread plugin by passing
comment->selection_type. but it doing so, i believe, there's a hole
in when setting the initial value to the comment's selection_type
field. well.